% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_rain_direct_and_evaporation.R
\name{add_rain_direct_and_evaporation}
\alias{add_rain_direct_and_evaporation}
\title{Add Direct Rain and Evaporation to flows_in_out config for sections with defined area_m2}
\usage{
add_rain_direct_and_evaporation(config)
}
\arguments{
\item{config}{config object as retrieved by \code{\link{config_read}}}
}
\value{
config with added flows_in_out for sections with defined area_m2, in
case none area_m2 is defined the original config is returned
}
\description{
Add Direct Rain and Evaporation to flows_in_out config for sections with defined area_m2
}
