% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_scenario.R
\name{add_scenario}
\alias{add_scenario}
\title{Add scenario}
\usage{
add_scenario(config, use_scenario = FALSE, debug = TRUE)
}
\arguments{
\item{config}{config as retrieved by \code{\link{config_read}}}

\item{use_scenario}{use scenario (default: FALSE)}

\item{debug}{print debug messages (default: TRUE)}
}
\value{
list with input parameters required for function \code{\link{prepare_input}}
}
\description{
Add scenario
}
