% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_substances.R
\name{add_substances}
\alias{add_substances}
\title{Add substances}
\usage{
add_substances(config)
}
\arguments{
\item{config}{config as retrieved by \code{\link{config_read}}}
}
\value{
adds new columns 'conc_<substance-name.kg.m3' and re-calculates input
concentrations from (ng|ug|mg|g)/L to kg/m3
}
\description{
Add substances
}
