% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_tracers.R
\name{add_tracers}
\alias{add_tracers}
\title{Add tracers}
\usage{
add_tracers(config)
}
\arguments{
\item{config}{config object as retrieved by \code{\link{config_read}}}
}
\value{
adds new columns 'conc_tracer.xxx' for four tracers ('CSO', 'inlet',
'rain_runoff' and 'WWTP') to flows_in_out for tracking sources withint the
water cycle
}
\description{
Add tracers
}
