% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_flows_monthly.R
\name{aggregate_flows_monthly}
\alias{aggregate_flows_monthly}
\title{Aggregate Flows Monthly}
\usage{
aggregate_flows_monthly(flows, aggregation_function = median)
}
\arguments{
\item{flows}{flows as retrieved by \code{\link{calculate_flows_auto}}}

\item{aggregation_function}{function used for aggregation (default: median)}
}
\value{
tibble with monthly aggregated flows data
}
\description{
Aggregate Flows Monthly
}
