% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_qualities_monthly.R
\name{aggregate_qualities_monthly}
\alias{aggregate_qualities_monthly}
\title{Aggregate Qualities Monthly}
\usage{
aggregate_qualities_monthly(
  qualities,
  aggregation_function = median,
  minimum_tracer_sum = 0.999
)
}
\arguments{
\item{qualities}{list as retrieved by \code{\link{calcualate_qualities}}}

\item{aggregation_function}{function used for aggregation (default: median)}

\item{minimum_tracer_sum}{minimum tracer sum (default: 0.999 i.e. 99.9 percent)
for filtering out results for with tracer has not reached almost 100 percent}
}
\value{
list with sublist "conc" with monthly aggregated concentrations data
}
\description{
Aggregate Qualities Monthly
}
