% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bankfiltration_shares_dynamic.R
\name{bfs_convert_equation}
\alias{bfs_convert_equation}
\title{Bank Filtration Share: convert equation}
\usage{
bfs_convert_equation(
  df,
  col_equation_a = "equation_a",
  col_equation_b = "equation_b"
)
}
\arguments{
\item{df}{data frame with bank filtration share equations, defined in sublist
config$bfstypes_equations as retrieved by \code{\link{config_read}}}

\item{col_equation_a}{column name of equation parameter a (default: "equation_a")}

\item{col_equation_b}{column name of equation parameter b (default: "equation_b")}
}
\value{
adds "equation_function" to data frame
}
\description{
Bank Filtration Share: convert equation
}
