% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_quality.R
\name{calc_conc}
\alias{calc_conc}
\title{Calculate Concentration}
\usage{
calc_conc(c_in, c_0, Q, V, k, t)
}
\arguments{
\item{c_in}{c_in concentration of inflow}

\item{c_0}{c_0 concentration in section at t = 0}

\item{Q}{total inflow rate into section (m3/s)}

\item{V}{volume of section (m3)}

\item{k}{degradation parameter}

\item{t}{time in seconds (s)}
}
\value{
concentration of substance at specific time
}
\description{
Calculate Concentration
}
\examples{
calc_conc(c_in = 10, c_0 = 0, Q = 40000, V = 300000, k = 0, t = 1)

}
