% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_flow.R
\name{calculate_flow}
\alias{calculate_flow}
\title{Calculate Flow}
\usage{
calculate_flow(
  df,
  input,
  shares_timeseries_wide = NULL,
  config,
  use_dynamic = FALSE,
  return_inputs = FALSE,
  debug = TRUE
)
}
\arguments{
\item{df}{data frame with model as retrieved by \code{\link{get_flowpath_table}}}

\item{input}{input flows as retrieved by \code{\link{prepare_input}} and sublist
"flows"}

\item{shares_timeseries_wide}{shares timeseries in wide formate, as retrieved
by \code{\link{prepare_input}} and sublist "shares_timeseries" (default: NULL),
only used if parameter "use_dynamic" is set to TRUE}

\item{config}{list with config as imported with \code{\link{config_read}}}

\item{use_dynamic}{for multiple outputs only: should static shares
(as defined in column "section_out_share" of "outflows_multiple.csv") be used
for separating the flow within a section or a function (as defined in column
"section_out_function" of "outflows_multiple.csv")), (default: FALSE)}

\item{return_inputs}{should also input data be returned in result dataset
(default: FALSE)}

\item{debug}{print debug messages (default: TRUE)}
}
\value{
tibble with Urban Water Model results
}
\description{
Calculate Flow
}
