% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{calculate_flow_share}
\alias{calculate_flow_share}
\title{Calculate flow share}
\usage{
calculate_flow_share(
  flow,
  shares_timeseries_id,
  shares_timeseries_wide,
  debug = FALSE
)
}
\arguments{
\item{flow}{vector with flows (same length as)}

\item{shares_timeseries_id}{id column in shares_timeseries_wide}

\item{shares_timeseries_wide}{tibble or data frame with shares time series
data in wide format}

\item{debug}{print debug messages (default: FALSE)}
}
\value{
flow vector corrected with values from shares_timeseries data.frame
for provided shares_timeseries_id
}
\description{
Calculate flow share
}
