% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_flowpath.R
\name{calculate_flowpath}
\alias{calculate_flowpath}
\title{Calculate Flowpath}
\usage{
calculate_flowpath(flow_name, links, nodes, backward = FALSE)
}
\arguments{
\item{flow_name}{flow name (as named in nodes) where calculation should start}

\item{links}{links}

\item{nodes}{nodes}

\item{backward}{should flowpath be calculated backwards? (default: FALSE)}
}
\value{
data frame with column "order" indicating the flow path order
}
\description{
Calculate Flowpath
}
