% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_flows_auto.R
\name{calculate_flows_auto}
\alias{calculate_flows_auto}
\title{Calculate Flows Automatically}
\usage{
calculate_flows_auto(
  config,
  input_list,
  network,
  use_dynamic = FALSE,
  debug = FALSE
)
}
\arguments{
\item{config}{list with config as imported with \code{\link{config_read}}}

\item{input_list}{input_list as retrieved by \code{\link{prepare_input}}}

\item{network}{tibble with water cycle flow network data, as retrieved by
\code{{\link{prepare_network}}}}

\item{use_dynamic}{for multiple outputs only: should static shares
(as defined in column "section_out_share" of "outflows_multiple.csv") be used
for separating the flow within a section or a function (as defined in column
"section_out_function" of "outflows_multiple.csv")), (default: FALSE)}

\item{debug}{print debug messages (default: FALSE)}
}
\value{
tibble with flows
}
\description{
Calculate Flows Automatically
}
