% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_outflow.R
\name{calculate_outflows}
\alias{calculate_outflows}
\title{Calculate Outflows}
\usage{
calculate_outflows(
  list_order,
  input,
  config,
  use_dynamic = TRUE,
  return_inputs
)
}
\arguments{
\item{list_order}{list for each order id (corresponding to same level of distance
of section from selected outflow)}

\item{input}{input as retrieved by \code{\link{prepare_input}}}

\item{config}{list with config as imported with \code{\link{config_read}}}

\item{use_dynamic}{for multiple outputs only: should static shares
(as defined in column "section_out_share" of "outflows_multiple.csv") be used
for separating the flow within a section or a function (as defined in column
"section_out_function" of "outflows_multiple.csv")), (default: TRUE)}

\item{return_inputs}{should also input data be returned in result dataset
(default: FALSE)}
}
\value{
tibble with outflows added to provided dataset
}
\description{
Calculate Outflows
}
