% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_qualities_backward.R
\name{calculate_qualities_backward}
\alias{calculate_qualities_backward}
\title{Calculate Qualities Backward}
\usage{
calculate_qualities_backward(input_list, flows, network, config, debug = TRUE)
}
\arguments{
\item{input_list}{input_list as retrieved by \code{\link{prepare_input}}}

\item{flows}{flows as retrieved by xxxx}

\item{network}{tibble with water cycle flow network data, as retrieved by
\code{{\link{prepare_network}}}}

\item{config}{list with config as imported with \code{\link{config_read}}}

\item{debug}{print debug messages (default: TRUE)}
}
\value{
returns modelled flows in tibble format
}
\description{
Calculate Qualities Backward
}
