% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_qualities_forward.R
\name{calculate_qualities_forward}
\alias{calculate_qualities_forward}
\title{Calculate Qualities Forward}
\usage{
calculate_qualities_forward(
  input_list,
  flows,
  network,
  config,
  max_sections = NULL,
  debug = TRUE
)
}
\arguments{
\item{input_list}{input_list as retrieved by \code{\link{prepare_input}}}

\item{flows}{flows as retrieved by xxxx}

\item{network}{tibble with water cycle flow network data, as retrieved by
\code{{\link{prepare_network}}}}

\item{config}{list with config as imported with \code{\link{config_read}}}

\item{debug}{print debug messages (default: TRUE)}

\item{max_section}{restrict number of calculated sections in case problem
occcur. Provide a number <= number of sections to be calculated. If NULL, all
sections will be calculated (default: NULL).}
}
\value{
returns modelled flows in tibble format
}
\description{
Calculate Qualities Forward
}
