% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_quality.R
\name{calculate_quality}
\alias{calculate_quality}
\title{Calculate Quality in Section}
\usage{
calculate_quality(
  df,
  input,
  shares_timeseries_wide = NULL,
  flows,
  quality = NULL,
  config,
  reverse_flow = FALSE,
  result_type = "list",
  debug = FALSE
)
}
\arguments{
\item{df}{data frame with model as retrieved by \code{\link{get_flowpath_table}}}

\item{input}{model input data flows as retrieved by \code{\link{prepare_input}}
and sublist "flows"}

\item{shares_timeseries_wide}{shares timeseries in wide format, as retrieved
by \code{\link{prepare_input}} and sublist "shares_timeseries" (default: NULL),
only used if parameter "use_dynamic" is set to TRUE}

\item{flows}{flows as retrieved by \code{\link{calculate_flows}}}

\item{quality}{list with sublist "conc" and "load" for all sections (default:
NULL)}

\item{config}{list with config as imported with \code{\link{config_read}}}

\item{reverse_flow}{calculate reverse flow (default: FALSE)}

\item{result_type}{define how the results should be returned. either a tibble
with loads and concentrations in long format.(if result_type == "tibble"), a list
with concentrations in wide format (result_type == "list") or a list with loads
in wide format (result_type == "load")}

\item{debug}{print debug messages (default: FALSE)}
}
\value{
tibble with Urban Water Model results with loads and concentrations
(if result_type == "raw"), a list with concentrations (result_type ==  "conc" or "load")
or a list with loads, default: "list"
}
\description{
Calculate Quality in Section
}
