% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_backflows-multiple.R
\name{check_backflows_multiple}
\alias{check_backflows_multiple}
\title{Check backflows multiple}
\usage{
check_backflows_multiple(config)
}
\arguments{
\item{config}{list with config as imported with \code{\link{config_read}}}
}
\value{
nothing besides stops the code in case checks do not pass or message
in case no config file configs/backflows_multiple.csv is provided
}
\description{
Check backflows multiple
}
