% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{check_if_function_and_calculate_flow}
\alias{check_if_function_and_calculate_flow}
\title{Check if function and calculate flow}
\usage{
check_if_function_and_calculate_flow(df, q)
}
\arguments{
\item{df}{data frame with one row containing column 'section_out_function_parsed'
with function for calculating flow}

\item{q}{flow vector}
}
\value{
if function: calculated flow vector, if not: NA_real
}
\description{
Check if function and calculate flow
}
