% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_network_errors.R
\name{check_network_errors}
\alias{check_network_errors}
\title{Check Network Errors}
\usage{
check_network_errors(network)
}
\arguments{
\item{network}{tibble with water cycle flow network network, as retrieved by
\code{{\link{prepare_network}}}}
}
\value{
list with information on network errors
}
\description{
Check Network Errors
}
