% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_outflow-multiple_dynamic-functions.R
\name{check_outflow_multiple_dynamic_functions}
\alias{check_outflow_multiple_dynamic_functions}
\title{Check outflow multiple dynamic functions}
\usage{
check_outflow_multiple_dynamic_functions(
  config,
  q = 1,
  allowed_relative_offset_percent = 0.001
)
}
\arguments{
\item{config}{list with config as imported with \code{\link{config_read}}}

\item{q}{total section flow used for testing, needs to be a scalar! (default: 1)}

\item{allowed_relative_offset_percent}{maximum allowed percental offset for sum of all
section outflows compared to section inflow (i.e. parameter q), default: 0.0001)}
}
\value{
nothingt if check passes for all outflow_id s with defined functions,
otherwise error
}
\description{
Check outflow multiple dynamic functions
}
