% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_shares_timeseries.R
\name{check_shares_timeseries}
\alias{check_shares_timeseries}
\title{Check Shares of Multiple Outflow Time Series}
\usage{
check_shares_timeseries(shares_timeseries, config, debug = TRUE)
}
\arguments{
\item{shares_timeseries}{shares timeseries dataset (default: kwb.BerlinWaterModel.public::shares_timeseries)}

\item{config}{model network configuration (as retrieved by \code{\link{config_read}})}

\item{debug}{print debug messages (default: TRUE)}
}
\value{
prints debug messages or stop in case of errors
}
\description{
Check Shares of Multiple Outflow Time Series
}
\examples{
 config <- kwb.BerlinWaterModel.public::config_read()
 check_shares_timeseries(
 shares_timeseries = kwb.BerlinWaterModel.public::shares_timeseries,
 config = config
 )
}
