% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_and_clean_dfs.R
\name{combine_and_clean_dfs}
\alias{combine_and_clean_dfs}
\title{Combine and clean dataframe}
\usage{
combine_and_clean_dfs(list)
}
\arguments{
\item{list}{list miwth Urban Water Model results}
}
\value{
tibble with Urban Water Model results
}
\description{
Combine and clean dataframe
}
