% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_read.R
\name{config_read}
\alias{config_read}
\title{Configuration: Read}
\usage{
config_read(
  config_dir = system.file("extdata/config/network_complete", package =
    "kwb.BerlinWaterModel.public"),
  file_encoding = "UTF-8"
)
}
\arguments{
\item{config_dir}{directory with configuration files (default:
system.file("extdata/config/network_complete", package = "kwb.BerlinWaterModel")). It is mandatory
that there are three files within this folder: "flows_in_out.csv", "outflows_multiple.csv"
and "sections.csv"}

\item{file_encoding}{encoding for reading the files (default: "UTF-8")}
}
\value{
list with three sublists "flows_in_out", "outflows_multiple" and
"sections
}
\description{
Configuration: Read
}
\examples{
\dontrun{config <- kwb.BerlinWaterModel.public::config_read()
config}
}
