% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_concentration_units.R
\name{convert_concentration_units}
\alias{convert_concentration_units}
\title{Convert concentration units to kg/m3}
\usage{
convert_concentration_units(df, return_inputs = FALSE)
}
\arguments{
\item{df}{data frame with "conc_\if{html}{\out{<substance-short-name>}}.<g|mg|mg|ng>.\if{html}{\out{<L>}}"}

\item{return_inputs}{should input data frame also be returned (default: FALSE)
or only newly calculated columns?}
}
\value{
tibble with "conc_\if{html}{\out{<substance-short-name>}}.kg.m3"
}
\description{
Convert concentration units to kg/m3
}
\examples{
df <- data.frame(
conc_As.ng.L = c(10, 50, 100),  # Nanogramm pro Liter
conc_Zn.ug.L = c(500, 1000, 2000),  # Mikrogramm pro Liter
conc_Fe.mg.L = c(0.1, 0.2, 0.3),    # Milligramm pro Liter
conc_Cu.g.L = c(0.005, 0.01, 0.02)  # Gramm pro Liter
)

convert_concentration_units(df, return_inputs = TRUE)

}
