% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_cso.R
\docType{data}
\name{cso}
\alias{cso}
\title{Dataset: Combined Sewer Overflows (cso) for one rainfall event simulated with Infoworks in 2019}
\format{
A tibble with 12,096 rows and five columns
\describe{
\item{file_name}{filename of data origin}
\item{datetime}{datetime}
\item{seconds}{seconds}
\item{CSO_id}{CSO_id}
\item{cbm_per_second}{flow (cbm per second)}
}
}
\usage{
cso
}
\description{
Dataset with combined sewer overflow events for one rainfall rainfall event simulated with Infoworks
in 2019
}
\keyword{datasets}
