% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_timeseries.R
\name{fill_timeseries}
\alias{fill_timeseries}
\title{Fill Timeseries based on User Defined Intervall (days, hours, minutes or seconds)}
\usage{
fill_timeseries(
  df,
  col_datetime = "date",
  temporal_resolution = "days",
  direction = "up"
)
}
\arguments{
\item{df}{data frame with data}

\item{col_datetime}{column of date or datetime}

\item{temporal_resolution}{select one of ("days", "hours", "minutes", "seconds"),
(default: "days")}

\item{direction}{in which direction should the parameter be filled ("up" or
"down"), default: "up"}
}
\value{
tibble with filled dates. Note that if min(daste) == "xxxx-xx-end-of-month-day"
the min(date) is set to "xxxx-xx-01", as monthly measured (originally summed)
values were set to the last day of the month, but are valid for the whole month
}
\description{
Fill Timeseries based on User Defined Intervall (days, hours, minutes or seconds)
}
