% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_flowpath.R
\name{find_node_orders}
\alias{find_node_orders}
\title{Helper function: find node order}
\usage{
find_node_orders(links, nodes, start_node, reverse = FALSE, mode = "out")
}
\arguments{
\item{links}{links}

\item{nodes}{nodes}

\item{start_node}{start node}

\item{reverse}{should network direction be reversed? (default: FALSE)}

\item{mode}{Character constant, gives whether the shortest paths to or from
the given vertices should be calculated for directed graphs. If out then the
shortest paths from the vertex, if ⁠in⁠ then to it will be considered. If all,
the default, then the graph is treated as undirected, i.e. edge directions
are not taken into account. This argument is ignored for undirected graphs
(default: "out")}
}
\value{
data frame with additional column "order"
}
\description{
Helper function: find node order
}
