% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bfshares.R
\name{get_bfshares}
\alias{get_bfshares}
\title{Get Bank Filtration shares Model input}
\usage{
get_bfshares(
  config,
  ww = kwb.BerlinWaterModel::ww,
  temporal_resolution = "days",
  bfshare_dynamic = TRUE
)
}
\arguments{
\item{config}{model network configuration (as retrieved by \code{\link{config_read}})}

\item{ww}{waterworks dataset (default: kwb.BerlinWaterModel::ww)}

\item{temporal_resolution}{specify temporal resolution of model input dataset.
(default: "days"). Valid options are: "days" or "hours"}

\item{bfshare_dynamic}{should dynamic bankfiltration shares be used or the static
ones contained in column "bank_filtration_share" of config$flows_in_out}
}
\value{
data framer with flow ids, date/datetime and well gallery metadata,
column "bank_filtration_share" is set depending on bfshare_dynamic == TRUE
(based on column "bank_filtration_share_dynamic") or FALSE (bank_filtration_share_static )
}
\description{
Get Bank Filtration shares Model input
}
