% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_date_or_datetime_columns}
\alias{get_date_or_datetime_columns}
\title{Helper Get date or datetime columns}
\usage{
get_date_or_datetime_columns(df)
}
\arguments{
\item{df}{data frame}
}
\value{
names of columns of class Date or POSIXct
}
\description{
Helper Get date or datetime columns
}
\keyword{internal}
