% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_flowpath_table.R
\name{get_flowpath_table}
\alias{get_flowpath_table}
\title{Get Flowpath Table}
\usage{
get_flowpath_table(outflow_id, network, config)
}
\arguments{
\item{outflow_id}{id of outflow section}

\item{network}{tibble with water cycle flow network data, as retrieved by
\code{{\link{prepare_network}}}}

\item{config}{config}
}
\value{
list for each order id containing tibble with all sections of same
order id
}
\description{
Get Flowpath Table
}
