% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_ids_from_names}
\alias{get_ids_from_names}
\title{Helper function: get ids from names}
\usage{
get_ids_from_names(vector, config)
}
\arguments{
\item{vector}{vector with names}

\item{config}{config as retrieved by \code{\link{config_read}}}
}
\value{
tibble with section names and ids contained in configuration
}
\description{
Helper function: get ids from names
}
