% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_link_names}
\alias{get_link_names}
\title{Helper function: get link names}
\usage{
get_link_names(network)
}
\arguments{
\item{network}{tibble with water cycle flow network data, as retrieved by
\code{{\link{prepare_network}}}}
}
\value{
tibble with link names and columns source and target
}
\description{
Helper function: get link names
}
