% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_names_from_ids}
\alias{get_names_from_ids}
\title{Helper function: get names from ids}
\usage{
get_names_from_ids(vector, config)
}
\arguments{
\item{vector}{vector with ids}

\item{config}{config as retrieved by \code{\link{config_read}}}
}
\value{
tibble with section names and ids contained in configuration
}
\description{
Helper function: get names from ids
}
