% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_nodes}
\alias{get_nodes}
\title{Helper function: get nodes}
\usage{
get_nodes(network, config)
}
\arguments{
\item{network}{tibble with water cycle flow network data, as retrieved by
\code{{\link{prepare_network}}}}

\item{config}{config as retrieved by \code{\link{config_read}}}
}
\value{
tibble with columns name (from_name) and group (source_group)
}
\description{
Helper function: get nodes
}
