% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_reverse_flows_per_section}
\alias{get_reverse_flows_per_section}
\title{Get Reverse Flows Per Section}
\usage{
get_reverse_flows_per_section(flows)
}
\arguments{
\item{flows}{flows}
}
\value{
tibble with sections ordered decreasing by number of data points below
zero and date_min (first neg. flow) and date_max (last neg. flow) within dataset
}
\description{
Get Reverse Flows Per Section
}
