% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_section_idnames}
\alias{get_section_idnames}
\title{Helper function: get section idnames}
\usage{
get_section_idnames(config)
}
\arguments{
\item{config}{config as retrieved by \code{\link{config_read}}}
}
\value{
tibble with section names and ids contained in configuration
}
\description{
Helper function: get section idnames
}
