% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_inflows.R
\docType{data}
\name{inflows}
\alias{inflows}
\title{Dataset: Inflows}
\format{
A tibble with 398,840 rows and three columns
\describe{
\item{date}{date}
\item{id}{id of inflow station}
\item{cbm_per_second}{measured flow in cubicmeter per second}
}
}
\usage{
inflows
}
\description{
Dataset with inflows
}
\keyword{datasets}
