% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bankfiltration_shares_dynamic.R
\name{parse_a_and_b}
\alias{parse_a_and_b}
\title{Helper function: parse a and}
\usage{
parse_a_and_b(a, b)
}
\arguments{
\item{a}{parameter a}

\item{b}{parameter b}
}
\value{
R function of parsed equation
}
\description{
Helper function: parse a and
}
\examples{
a <- 1
b <- 2
parse_a_and_b(a, b)
}
