% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bankfiltration_shares_dynamic.R
\name{parse_equation}
\alias{parse_equation}
\title{Helper function: parse equation}
\usage{
parse_equation(equation)
}
\arguments{
\item{equation}{equation in text form (either "constant" -> y = c or "ln" ->
y = a * ln(x) + b)}
}
\value{
R function of parsed equation
}
\description{
Helper function: parse equation
}
