% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_flows_and_bfshares_per_section.R
\name{plot_flows_and_bfshares_per_section}
\alias{plot_flows_and_bfshares_per_section}
\title{Plot Flows and BF & MAR shares per section}
\usage{
plot_flows_and_bfshares_per_section(
  config,
  flows,
  network,
  ww,
  scale_factor = 1,
  debug = TRUE,
  add_caption_simulation = FALSE,
  add_caption_weblink = FALSE,
  add_caption_outflowsmultiple = FALSE
)
}
\arguments{
\item{config}{list with config as imported with \code{\link{config_read}}}

\item{flows}{flows as retrieved by xxxx}

\item{network}{tibble with water cycle flow network data, as retrieved by
\code{{\link{prepare_network}}}}

\item{ww}{waterworks dataset (default: kwb.BerlinWaterModel.public::ww)}

\item{scale_factor}{scale factor for increasing size of labels}

\item{debug}{print debug messages (default: TRUE)}

\item{add_caption_simulation}{add caption with simulation information (default: FALSE)}

\item{add_caption_weblink}{Add one line in caption to reference webpage of vignette (default: FALSE)}

\item{add_caption_outflowsmultiple}{Add information of model configuration for
multiple outflows (default: FALSE)}
}
\value{
plots flows and BF & MAR shares for each section
}
\description{
Plot Flows and BF & MAR shares per section
}
