% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{plot_network_complex}
\alias{plot_network_complex}
\title{Plot Network: complex}
\usage{
plot_network_complex(network, config, show_labels = FALSE)
}
\arguments{
\item{network}{tibble with water cycle flow network data, as retrieved by
\code{{\link{prepare_network}}}}

\item{config}{list with config as imported with \code{\link{config_read}}}

\item{show_labels}{show labels (default: FALSE)}
}
\value{
complex network graph
}
\description{
Plot Network: complex
}
