% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{plot_network_simple}
\alias{plot_network_simple}
\title{Plot Network: simple}
\usage{
plot_network_simple(network)
}
\arguments{
\item{network}{tibble with water cycle flow network data, as retrieved by
\code{{\link{prepare_network}}}}
}
\value{
simple network graph
}
\description{
Plot Network: simple
}
