% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_network.R
\name{prepare_network}
\alias{prepare_network}
\title{Prepare Network Data}
\usage{
prepare_network(config)
}
\arguments{
\item{config}{config as retrieved by \code{\link{config_read}}}
}
\value{
data structure for plot_network functions and for further calculations
}
\description{
Prepare Network Data
}
