% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_qsimVis_input.R
\name{prepare_qsimVis_input}
\alias{prepare_qsimVis_input}
\title{Preparew QsimVis Input}
\usage{
prepare_qsimVis_input(config, flows, qualities, rounding_digits = 3)
}
\arguments{
\item{config}{config list as retrieved by \code{\link{config_read}}}

\item{flows}{flows input data frame in "wide" format}

\item{qualities}{qualities list input data structure for each section in wide format}

\item{rounding_digits}{digits used for result data rounding (default: 3)}
}
\value{
data frame with flow data in long format joined with
config$qsimVis data frame
}
\description{
Preparew QsimVis Input
}
