% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_rain.R
\docType{data}
\name{rain}
\alias{rain}
\title{Dataset: Rain (rain) with DWD rainfall data}
\format{
A tibble with 247,076 rows and two columns
\describe{
\item{datetime}{datetime}
\item{DWD_0433}{rainfall in mm for DWD rainstation 0433}
}
}
\usage{
rain
}
\description{
Dataset with rainfall data (based on DWD station 0433)
}
\keyword{datasets}
