% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_tracer_starting_conc.R
\name{set_tracer_starting_conc}
\alias{set_tracer_starting_conc}
\title{Set Tracer Starting Concentrations To Final Modelling Results}
\usage{
set_tracer_starting_conc(config, qualities)
}
\arguments{
\item{config}{config list structure as retrieved by \code{\link{config_read}}}

\item{qualities}{qualities list result structure}
}
\value{
config list with starting concentrations (defined in config$sections)
based on qualities. The last time step is used as starting concentration
}
\description{
Set Tracer Starting Concentrations To Final Modelling Results
}
