% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_tracer_starting_conc_stats.R
\name{set_tracer_starting_conc_stats}
\alias{set_tracer_starting_conc_stats}
\title{Set Tracer Starting Concentrations Statistics}
\usage{
set_tracer_starting_conc_stats(
  config,
  qualities,
  aggregation_function = median,
  minimum_tracer_sum = 0.999
)
}
\arguments{
\item{config}{config list structure as retrieved by \code{\link{config_read}}}

\item{qualities}{qualities list result structure}

\item{aggregation_function}{function used for data aggregation (default: median)}

\item{minimum_tracer_sum}{minimum tracer sum (default: 0.999 i.e. 99.9 percent)
for filtering out results for with tracer has not reached almost 100 percent}
}
\value{
config list with starting concentrations (defined in config$sections)
based on qualities
}
\description{
Set Tracer Starting Concentrations Statistics
}
