% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_shares_timeseries.R
\docType{data}
\name{shares_timeseries}
\alias{shares_timeseries}
\title{Dataset: Shares (0 - 1) of discharge to a specific downstream river section}
\format{
A tibble with 15,340 rows and three columns
\describe{
\item{date}{date}
\item{id}{id of shares timeseries}
\item{share}{share (0-1) to a specific downstream river section}
}
}
\usage{
shares_timeseries
}
\description{
Dataset: Shares (0 - 1) of discharge to a specific downstream river section
}
\keyword{datasets}
