% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{shorten_ww_flow_id}
\alias{shorten_ww_flow_id}
\title{Helper: shorten WW flow id}
\usage{
shorten_ww_flow_id(df, col_flow_id = "flow_id")
}
\arguments{
\item{df}{df with WW flow data andf flow ids as columns}

\item{col_flow_id}{column name with flow id (default: "flow_id")}
}
\value{
vector with shortened WW flow ids
}
\description{
Helper: shorten WW flow id
}
