% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_ww.R
\docType{data}
\name{ww}
\alias{ww}
\title{Dataset: Waterworks (WW) Abstractions per Gallery}
\format{
A tibble with 17,141 rows and three columns
\describe{
\item{date}{date}
\item{id}{id of WW gallery}
\item{cbm_per_second}{measured flow in cubicmeter per second, backcalculated
from monthly abstraction sums by dividing through number of days in month}
}
}
\usage{
ww
}
\description{
Dataset with waterworks abstractions per gallery
}
\keyword{datasets}
