% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_wwtp.R
\docType{data}
\name{wwtp}
\alias{wwtp}
\title{Dataset: WWTP (Wastewater Treatment Plant) Flows}
\format{
A tibble with 1,152 rows and three columns
\describe{
\item{date}{date}
\item{id}{id of WWTP}
\item{cbm_per_second}{measured flow in cubicmeter per second, backcalculated
from monthly effluent sums by dividing through number of days in month}
}
}
\usage{
wwtp
}
\description{
Dataset with WWTP inflows
}
\keyword{datasets}
