% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_qualities_forward.R
\name{calculate_qualities_forward}
\alias{calculate_qualities_forward}
\title{Calculate Qualities Forward}
\usage{
calculate_qualities_forward(
  input_list,
  flows,
  network,
  config,
  max_sections = NULL,
  debug = TRUE
)
}
\arguments{
\item{input_list}{input_list as retrieved by \code{\link{prepare_input}}}

\item{flows}{tibble with (modelled) flows for the network (e.g. precomputed
upstream by the user)}

\item{network}{tibble with water cycle flow network data, as retrieved by
\code{\link{prepare_network}}}

\item{config}{list with config as imported with \code{\link{config_read}}}

\item{max_sections}{restrict number of calculated sections in case problems
occur. Provide a number <= number of sections to be calculated. If NULL, all
sections will be calculated (default: NULL). Only used if reverse_flow = FALSE}

\item{debug}{print debug messages (default: TRUE)}
}
\value{
returns modelled flows in tibble format
}
\description{
Calculate Qualities Forward
}
